Updates color choices and changes layouts
