+++ link-admin/package.json
@@ -23,6 +23,7 @@
     "@zendeskgarden/react-theming": "^8.21.2",
     "@zendeskgarden/react-typography": "^8.21.2",
     "@zendeskgarden/svg-icons": "^6.27.0",
+    "@zendeskgarden/tailwindcss": "^1.1.0",
     "axios": "^0.20.0",
     "classnames": "^2.2.6",
     "connected-react-router": "^6.8.0",
@@ -32,6 +33,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-hook-form": "^6.14.0",
     "react-redux": "^7.2.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
@@ -48,7 +50,6 @@
     "@storybook/addons": "^6.0.21",
     "@storybook/preset-create-react-app": "^3.1.4",
     "@storybook/react": "^6.0.21",
-    "@zendeskgarden/tailwindcss": "^1.1.0",
     "autoprefixer": "^9.8.6",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.4",
