+++ link-admin/src/components/Taxonomy/index.js
@@ -1,0 +1,3 @@
+import Tag from './Tag'
+
+export { Tag }
